agentic_backend_pinecone/requirements.txt
# --- Core web stack ---
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
sse-starlette>=2.1.3
python-multipart>=0.0.9

# --- Config ---
pydantic>=2.7.0
python-dotenv>=1.0.0
pydantic-settings>=2.3.0

# --- LangChain + providers ---
langchain>=0.2.16
langchain-core>=0.2.43
langchain-community>=0.2.18
langchain-google-genai>=2.1.9
langchain-ollama>=0.2.0

# --- Pinecone integration ---
pinecone>=5.0.0           # modern SDK package name
langchain-pinecone>=0.1.1

# --- Embeddings ---
sentence-transformers>=3.0.1
numpy>=1.26.4
aiohttp>=3.9.5

# --- Vision ---
transformers>=4.44.2
Pillow>=10.3.0
torch>=2.3.0; platform_system != "Darwin"  # macOS users: install torch separately

# --- Audio utils ---
pydub>=0.25.1
